requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
dbt-core
dbt-postgres
pandas
pytest
